GB - Алгоритмы. Урок 3 Задание 1 - комменты и методы вычисления расстояния.
